CMake builds with all compiler warnings enabled and without outputting warnings
CMake now prints the current Build Type upon building

Project now compiles with no warnings invoked by the compiler.
Warnings will now remain enabled at all times.
No warnings tested with gcc version 11.3.0 and clang version 14.0.0 on an ubuntu 22.04 x86 system

set CMAKE_CXX_FLAGS_RELEASE in CMakeLists.txt to be "-O3"

Removed what() overload from ascii::ffmpeg_error

+++ src/ascii.cpp
@@ -74,7 +74,7 @@ char get_char_from_value(std::string& characters, uint8_t value) {
 }
 
 RGBColor get_rgb_from_char(std::string& characters, char ch) {
-    for (int i = 0; i < characters.size(); i++) {
+    for (int i = 0; i < (int)characters.size(); i++) {
         if (ch == characters[i]) {
             uint8_t val = i * 255 / (characters.size() - 1); 
             return RGBColor(val);
